separated comparable tests out.
