+++ Tests/AttributesTests/validators/ValidatorFactoryTests.swift
@@ -285,145 +285,6 @@ final class ValidationFactoryTests: XCTestCase {
         }
     }
 
-    /// Test between method.
-    func testBetween() throws {
-        let readPath = ReadOnlyPath(Int.self)
-        let factory = ValidatorFactory<Int>.required().between(min: 3, max: 7)
-        try factory.make(path: readPath).performValidation(5)
-        try factory.make(path: readPath).performValidation(3)
-        try factory.make(path: readPath).performValidation(7)
-        XCTAssertThrowsError(try factory.make(path: readPath).performValidation(1)) {
-            guard let error = $0 as? ValidationError<Int> else {
-                XCTFail("Incorrect error thrown.")
-                return
-            }
-            XCTAssertEqual(error.message, "Must be between \(3) and \(7).")
-            XCTAssertEqual(error.path, AnyPath(readPath))
-        }
-    }
-
-    /// Test `lessThan` method.
-    func testLessThan() throws {
-        let readPath = ReadOnlyPath(Int.self)
-        let factory = ValidatorFactory<Int>.required().lessThan(5)
-        try factory.make(path: readPath).performValidation(4)
-        XCTAssertThrowsError(try factory.make(path: readPath).performValidation(5)) {
-            guard let error = $0 as? ValidationError<Int> else {
-                XCTFail("Incorrect error thrown.")
-                return
-            }
-            XCTAssertEqual(error.message, "Must be less than \(5).")
-            XCTAssertEqual(error.path, AnyPath(readPath))
-        }
-    }
-
-    /// Test `lessThanEqual` method.
-    func testLessThanEqual() throws {
-        let readPath = ReadOnlyPath(Int.self)
-        let factory = ValidatorFactory<Int>.required().lessThanEqual(5)
-        try factory.make(path: readPath).performValidation(4)
-        try factory.make(path: readPath).performValidation(5)
-        XCTAssertThrowsError(try factory.make(path: readPath).performValidation(6)) {
-            guard let error = $0 as? ValidationError<Int> else {
-                XCTFail("Incorrect error thrown.")
-                return
-            }
-            XCTAssertEqual(error.message, "Must be less than or equal to \(5).")
-            XCTAssertEqual(error.path, AnyPath(readPath))
-        }
-    }
-
-    /// Test `greaterThan` method.
-    func testGreaterThan() throws {
-        let readPath = ReadOnlyPath(Int.self)
-        let factory = ValidatorFactory<Int>.required().greaterThan(5)
-        try factory.make(path: readPath).performValidation(6)
-        XCTAssertThrowsError(try factory.make(path: readPath).performValidation(5)) {
-            guard let error = $0 as? ValidationError<Int> else {
-                XCTFail("Incorrect error thrown.")
-                return
-            }
-            XCTAssertEqual(error.message, "Must be greater than \(5).")
-            XCTAssertEqual(error.path, AnyPath(readPath))
-        }
-    }
-
-    /// Test `greaterThanEqual` method.
-    func testGreaterThanEqual() throws {
-        let readPath = ReadOnlyPath(Int.self)
-        let factory = ValidatorFactory<Int>.required().greaterThanEqual(5)
-        try factory.make(path: readPath).performValidation(5)
-        try factory.make(path: readPath).performValidation(6)
-        XCTAssertThrowsError(try factory.make(path: readPath).performValidation(4)) {
-            guard let error = $0 as? ValidationError<Int> else {
-                XCTFail("Incorrect error thrown.")
-                return
-            }
-            XCTAssertEqual(error.message, "Must be greater than or equal to \(5).")
-            XCTAssertEqual(error.path, AnyPath(readPath))
-        }
-    }
-
-    /// Test `empty` rule.
-    func testEmpty() throws {
-        let readPath = ReadOnlyPath([Int].self)
-        let factory = ValidatorFactory<[Int]>.required().empty()
-        try factory.make(path: readPath).performValidation([])
-        XCTAssertThrowsError(try factory.make(path: readPath).performValidation([1, 2, 3])) {
-            handleError($0, message: "Must be empty.")
-        }
-    }
-
-    /// Test `notEmpty` rule.
-    func testNotEmpty() throws {
-        let readPath = ReadOnlyPath([Int].self)
-        let factory = ValidatorFactory<[Int]>.required().notEmpty()
-        try factory.make(path: readPath).performValidation([1, 2, 3])
-        XCTAssertThrowsError(try factory.make(path: readPath).performValidation([])) {
-            handleError($0, message: "Cannot be empty.")
-        }
-    }
-
-    /// Test `length` rule.
-    func testLength() throws {
-        let readPath = ReadOnlyPath([Int].self)
-        let factory = ValidatorFactory<[Int]>.required().length(3)
-        try factory.make(path: readPath).performValidation([1, 2, 3])
-        XCTAssertThrowsError(try factory.make(path: readPath).performValidation([1, 2])) {
-            handleError($0, message: "Must have exactly 3 elements.")
-        }
-        XCTAssertThrowsError(try factory.make(path: readPath).performValidation([])) {
-            handleError($0, message: "Must have exactly 3 elements.")
-        }
-    }
-
-    /// Test `minLength` rule.
-    func testMinLength() throws {
-        let readPath = ReadOnlyPath([Int].self)
-        let factory = ValidatorFactory<[Int]>.required().minLength(2)
-        try factory.make(path: readPath).performValidation([1, 2])
-        try factory.make(path: readPath).performValidation([1, 2, 3])
-        XCTAssertThrowsError(try factory.make(path: readPath).performValidation([1])) {
-            handleError($0, message: "Must provide at least 2 values.")
-        }
-        XCTAssertThrowsError(try factory.make(path: readPath).performValidation([])) {
-            handleError($0, message: "Must provide at least 2 values.")
-        }
-    }
-
-    /// Test `maxLength` rule.
-    func testMaxLength() throws {
-        let readPath = ReadOnlyPath([Int].self)
-        let factory = ValidatorFactory<[Int]>.required().maxLength(3)
-        try factory.make(path: readPath).performValidation([])
-        try factory.make(path: readPath).performValidation([1])
-        try factory.make(path: readPath).performValidation([1, 2])
-        try factory.make(path: readPath).performValidation([1, 2, 3])
-        XCTAssertThrowsError(try factory.make(path: readPath).performValidation([1, 2, 3, 4])) {
-            handleError($0, message: "Must provide no more than 3 values.")
-        }
-    }
-
     /// Test alpha rule.
     func testAlpha() throws {
         let readPath = ReadOnlyPath(String.self)
@@ -667,22 +528,6 @@ final class ValidationFactoryTests: XCTestCase {
         _ error: Error, message: String, path: AnyPath<String> = AnyPath(ReadOnlyPath(String.self))
     ) {
         guard let error = error as? ValidationError<String> else {
-            XCTFail("Incorrect error thrown.")
-            return
-        }
-        XCTAssertEqual(error.message, message)
-        XCTAssertEqual(error.path, path)
-    }
-
-    /// Handle the case where the validator throws a validation error during a collection validation.
-    /// - Parameters:
-    ///   - error: The abstract Error.
-    ///   - message: The message contained within the ``ValidationError``.
-    ///   - path: The path contained within the ``ValidationError``
-    private func handleError(
-        _ error: Error, message: String, path: AnyPath<[Int]> = AnyPath(ReadOnlyPath([Int].self))
-    ) {
-        guard let error = error as? ValidationError<[Int]> else {
             XCTFail("Incorrect error thrown.")
             return
         }
